+++ src/Module/System/Update.php
@@ -221,6 +221,9 @@ class Update
         $update_string = "* Delete 'concerts_limit_past' and 'concerts_limit_future' database settings.<br/ > ";
         $version[]     = array('version' => '400023', 'description' => $update_string);
 
+        $update_string = "**IMPORTANT UPDATE NOTES**<br />" . "These columns will fill dynamically in the web UI but you should do a catalog 'add' as soon as possible to fill them.<br />It will take a while for large libraries but will help API and SubSonic clients.<br /><br />" . "* Add 'song_count', 'album_count' and 'album_group_count' to artist. <br />";
+        $version[]     = array('version' => '400024', 'description' => $update_string);
+
         return $version;
     }
 
@@ -1159,6 +1162,24 @@ class Update
 
         $sql = "DELETE FROM `preference` " .
             "WHERE `preference`.`name` IN ('concerts_limit_past', 'concerts_limit_future');";
+        $retval &= Dba::write($sql);
+
+        return $retval;
+    }
+
+    /**
+     * update_400024
+     *
+     * Add song_count, album_count and album_group_count to artist
+     */
+    public static function update_400024()
+    {
+        $retval = true;
+        $sql    = "ALTER TABLE `artist` ADD `song_count` smallint(5) unsigned DEFAULT 0 NULL;";
+        $retval &= Dba::write($sql);
+        $sql    = "ALTER TABLE `artist` ADD `album_count` smallint(5) unsigned DEFAULT 0 NULL;";
+        $retval &= Dba::write($sql);
+        $sql    = "ALTER TABLE `artist` ADD `album_group_count` smallint(5) unsigned DEFAULT 0 NULL;";
         $retval &= Dba::write($sql);
 
         return $retval;